Add remote file system support for Utils
