+++ zoo/src/main/scala/com/intel/analytics/zoo/examples/streaming/objectdetection/StreamingObjectDetection.scala
@@ -18,7 +18,7 @@ package com.intel.analytics.zoo.examples.streaming.objectdetection
 
 
 import com.intel.analytics.bigdl.transform.vision.image.ImageFeature
-import com.intel.analytics.zoo.common.NNContext
+import com.intel.analytics.zoo.common.{NNContext, Utils}
 import com.intel.analytics.zoo.feature.image.{ImageBytesToMat, ImageSet}
 import com.intel.analytics.zoo.models.image.objectdetection.{ObjectDetector, Visualizer}
 import org.apache.hadoop.conf.Configuration
@@ -98,14 +98,7 @@ object StreamingObjectDetection {
    */
   def readFile(path: String): ImageFeature = {
     logger.info("Read image file " + path)
-    val fspath = new Path(path)
-    val fs = FileSystem.get(fspath.toUri, new Configuration())
-    // Read local or remote image
-    val inputStream = fs.open(fspath)
-    val data = new Array[Byte](fs.getFileStatus(new Path(path))
-      .getLen.toInt)
-    inputStream.readFully(data)
-    inputStream.close()
+    val data = Utils.readBytes(path)
     ImageFeature.apply(data, null, path)
   }
 
@@ -118,12 +111,7 @@ object StreamingObjectDetection {
   def writeFile(outPath: String, path: String, content: Array[Byte]): Unit = {
     val fspath = getOutPath(outPath, path, "jpg")
     logger.info("Writing image file " + fspath.toString)
-    val fs = FileSystem.get(fspath.toUri, new Configuration())
-    val outStream = fs.create(
-      fspath,
-      true)
-    outStream.write(content)
-    outStream.close()
+    Utils.saveBytes(content, fspath.toString, true)
   }
 
   /**
